+++ ui/js/address.js
@@ -1,2 +1,2 @@
 ADDRESS = 'localhost'
-PORT = '8888'
+PORT = '9000'