+++ ui/js/core.js
@@ -1,3 +1,7 @@
+var TABS = ['knowledge', 'thoughts']
+
+var loc = window.location.href.split('/')[3]
+var CURRENT_TAB = TABS.includes(loc) ? loc : 'knowledge';
 
 var colors = d3.scaleOrdinal(d3.schemeCategory10);
 var converter = new showdown.Converter();
@@ -27,6 +31,8 @@ function make_table(data, localstorage){
     for (var i = 0; i < data.length; i++){
         var item = data[i];
         var tags = gen_tags(item['tags']);
+        var tab = item['tab'];
+        if (tab != CURRENT_TAB) continue;
         html += item_template({
             key: converter.makeHtml(item['key']),
             val: converter.makeHtml(parse(item['val'])),
@@ -47,9 +53,13 @@ function get_local_items(){
     return local_items;
 }
 
+function tab_filter(item){
+    return item['tab'] == CURRENT_TAB;
+}
+
 function stats(data, local_items){
     return stats_template({
-        items: data['items'].length + local_items.length,
+        items: data['items'].filter(tab_filter).length + local_items.filter(tab_filter).length,
         tags: data['tags'].length,
     })
 }
@@ -68,9 +78,32 @@ function refresh(){
    get('/get?', _refresh);
 };
 
+function knowledge_tab(){
+    CURRENT_TAB = 'knowledge';
+    window.location.href = CURRENT_TAB;
+}
+function thoughts_tab(){
+    CURRENT_TAB = 'thoughts';
+    window.location.href = CURRENT_TAB;
+}
+
+function set_active_tab(){
+    $('.nav-item').each(function(){
+        $(this).removeClass('active');
+    });
+    $('#' + CURRENT_TAB + '-nav').attr('class', 'active');
+}
+
+
 $(document).ready(function(){
+    $('#add_inputs').html(add_template());
+    $('#table').html(table_template());
     refresh();
     $('#add').click(add);
+    $('#knowledge-nav').click(knowledge_tab);
+    $('#thoughts-nav').click(thoughts_tab);
+    set_active_tab();
+
     var title = window.location.href.split('/')[2];
     $(document).attr("title", 'brain-extension (' + title + ')');
 });