+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2018/09/12
 - Added tags recording in separate file and showed stats on ui.
+- Added separate tabs for Knowledge and Thoughts and stubs for todo and tags labels
 
 ## 2018/09/09
 - Added support for password